requirements.txt
python-telegram-bot==21.6
aiohttp==3.10.5
beautifulsoup4==4.12.3
feedparser==6.0.11
APScheduler==3.10.4
lxml==5.3.0
arxiv==2.1.3
requests>=2.32.0
pdfminer.six==20221105
openai==1.51.0
